pyproject.toml
[project]
name = "lease-data-foundation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib==3.10.8",
    "postal>=1.1.11",
    "protobuf==6.33.2",
    "psycopg2-binary>=2.9.11",
    "pymongo>=4.16.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "seaborn==0.13.2",
    "sentencepiece==0.2.1",
    "torch==2.9.1",
    "tqdm==4.67.1",
    "transformers==4.57.3",
]
